requirements.txt
flask
python-dotenv
google-genai
flask-login
pymongo
werkzeug
